feat(Planners.Core): Add SemanticMemoryConfig class.
Introduce a new SemanticMemoryConfig class to configure semantic memory settings for filtering function lookup during plan creation. This includes properties for IncludedFunctions, Memory, MaxRelevantFunctions, and RelevancyThreshold.
